+++ src/shaders/symbol_sdf.vertex.glsl
@@ -4,8 +4,8 @@ attribute vec4 a_pos_offset;
 attribute vec2 a_texture_pos;
 attribute vec4 a_data;
 
-#pragma mapbox: define lowp vec4 fill_color
-#pragma mapbox: define lowp vec4 halo_color
+#pragma mapbox: define highp vec4 fill_color
+#pragma mapbox: define highp vec4 halo_color
 #pragma mapbox: define lowp float opacity
 #pragma mapbox: define lowp float halo_width
 #pragma mapbox: define lowp float halo_blur
@@ -28,8 +28,8 @@ varying vec2 v_fade_tex;
 varying float v_gamma_scale;
 
 void main() {
-    #pragma mapbox: initialize lowp vec4 fill_color
-    #pragma mapbox: initialize lowp vec4 halo_color
+    #pragma mapbox: initialize highp vec4 fill_color
+    #pragma mapbox: initialize highp vec4 halo_color
     #pragma mapbox: initialize lowp float opacity
     #pragma mapbox: initialize lowp float halo_width
     #pragma mapbox: initialize lowp float halo_blur
